// src/app/modules/accounting/models/ye-general-journal-transaction.model.ts
export class YeGeneralJounalTrnDepositTrnModel{
    GLAccDesc: string;
    GLAccNo: number;
    GLCreditAmt: number;
    GLDebitAmt: number;
    Id: number;
    MarkRecord: number;
    TrnDesc: string; 
    TrnTypeDesc:string;
}

export class GLCatagoryModel{
    GLAccDesc: string;
    GLAccNo: number;
}

export class YeGeneralJournalTrnSavedData{
    GLAccDesc: string;
    GLAccNo: number;
    GLCreditAmt: number;
    GLDebitAmt: number;
    Id: number;
    TrnDesc: string;
    TrnTypeDes: string;
    VchNo: string;
}